+++ TP3 C++ /src/Algorithm/strategies/AlgorithmQuadra.h
@@ -12,7 +12,7 @@ class AlgorithmQuadra : public Algorithm {
 public:
     AlgorithmQuadra(const std::string &file);
 
-    int run() override;
+    long run() override;
 
 };
